Create  Pollution.js
Scope pollution is when we have too many global variables that exist in the global namespace, or when we reuse variables across different scopes. Scope pollution makes it difficult to keep track of our different variables and sets us up for potential accidents. 
While it's important to know what global scope is, it's best practice to not define variables in the global scope.